COS-5625: Remove ability to configure message for send connection request action
